basic error handling / logging
errors are now logged with default log library. I wrote a function that lets me be pretty lazy with them, as it prints out info about where it happened, but only if the error its passed isn't nil - I can sprinkle calls across the program, and it will handle the rest. I doubt its great for performance, but if errors are rare it should be fine. I would like to inform the user of errors, and write the error file for recovery, but I couldn't think of a good way to do either that wouldn't cause new errors sometimes.