display drive letter on windows
Fixes #2 by setting the full path to be the short name on windows, as the full path for drives will be a letter and a slash. This is not a great fix, and long term I intend to use win32 api and build flag to actually get drive names on windows. Also added some logging of GOOS and GOARCH, as i foresee issues with this solution.